requirements.txt
fastapi
uvicorn
logging
requests
PyPDF2
nltk
langchain
langchain_google_genai
langchain_groq
python-dotenv
faiss-cpu
pinecone
pinecone-client 
pinecone-text 
pinecone-notebooks